replace inject with store use
