+++ apps/web/src/components/identification/TeeRegisterHighlight.vue
@@ -16,7 +16,7 @@
             <TeeDsfrButton
               class="fr-text--yellow fr-text--bold"
               label="Je complète mon profil"
-              @click="toggleRegisterModal"
+              @click="openModal"
             />
           </div>
         </div>
@@ -25,9 +25,13 @@
   </DsfrHighlight>
 </template>
 <script setup lang="ts">
+import Navigation from '@/utils/navigation'
+
 interface Props {
   text: string
 }
 defineProps<Props>()
-const toggleRegisterModal = inject<() => void>('toggleRegisterModal')
+const openModal = () => {
+  Navigation.toggleRegisterModal()
+}
 </script>
