+++ apps/web/src/stores/navigation.ts
@@ -25,6 +25,7 @@ export const useNavigationStore = defineStore('navigation', () => {
   const stringOfSearchParams = ref<string>('')
   const tabSelectedOnList = ref<number>(0)
   const hasSpinner = ref<boolean>(false)
+  const hasRegisterModal = ref<boolean>(false)
   const query = computed<Record<string, LocationQueryValue | LocationQueryValue[]>>(() => {
     const query: LocationQuery = {}
     for (const key of new URLSearchParams(stringOfSearchParams.value).keys()) {
@@ -217,6 +218,7 @@ export const useNavigationStore = defineStore('navigation', () => {
     searchParams,
     tabSelectedOnList,
     hasSpinner,
+    hasRegisterModal,
     isCatalog,
     isCatalogAboutPrograms,
     isCatalogAboutProjects,
